refactor(vcs): move git service from sero-git-plugin into the host
Step 3 (part 1) of the VCS unification (AD-024). Fixes the inverted plugin
boundary: electron main no longer imports plugin code.

- extension/git-* + state-io move to features/vcs/git-service/ (with tests)
- GitAppState + subtypes/helpers move to @sero-ai/common (git-app.ts); the
  plugin's shared/types re-exports them so plugin UI code is untouched
- new GitServiceBridge in @sero-ai/common (globalThis-keyed, same pattern
  as the user-feedback bus): the host registers the service at startup and
  the plugin's git_manager tool + session_start hook delegate through it —
  actions on registered workspaces flow through GitWorkspaceStateManager so
  refresh invalidation applies
- runGitAsync (fetch/pull/push seam) now injects Sero's GitHub auth env,
  closing the token-injection gap for Git-app network operations

Plan: docs/features/vcs-unification.md (Phase 3, step 3)
