+++ apps/desktop/electron/features/vcs/git-service/git-default-branch.ts
