+++ apps/desktop/electron/features/vcs/git-service/git-refresh.ts
@@ -3,7 +3,7 @@ import type {
   GitAppState,
   GitSyncMode,
   RemoteInfo,
-} from '../shared/types';
+} from '@sero-ai/common';
 import {
   getCurrentBranch,
   getFileChanges,
